Move RestConfiguration to ignite-network
