+++ modules/network/src/main/java/org/apache/ignite/internal/network/configuration/RestConfigurationSchema.java
@@ -15,8 +15,26 @@
  * limitations under the License.
  */
 
-/**
- * Configuration schemas for Rest component.
- */
+package org.apache.ignite.internal.network.configuration;
 
-package org.apache.ignite.internal.rest.configuration;
+import org.apache.ignite.configuration.annotation.ConfigurationRoot;
+import org.apache.ignite.configuration.annotation.ConfigurationType;
+import org.apache.ignite.configuration.annotation.Value;
+import org.apache.ignite.configuration.validation.Range;
+
+/**
+ * Configuration schema for REST endpoint subtree.
+ */
+@SuppressWarnings("PMD.UnusedPrivateField")
+@ConfigurationRoot(rootName = "rest", type = ConfigurationType.LOCAL)
+public class RestConfigurationSchema {
+    /** TCP port. */
+    @Range(min = 1024, max = 0xFFFF)
+    @Value(hasDefault = true)
+    public final int port = 10300;
+
+    /** TCP port range. */
+    @Range(min = 0)
+    @Value(hasDefault = true)
+    public final int portRange = 100;
+}
